pyproject.toml

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "kubernetes_wrapper"
version = "0.2.1"
description = "Serialized Data Interface for Juju Operators"
authors = [
    "Dominik Fleischmann <dominik.fleischmann@canonical.com>",
    "Kenneth Koski <kenneth.koski@canonical.com>"
]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/canonical/kubernetes-wrapper/"
repository = "https://github.com/canonical/kubernetes-wrapper/"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License"
]

[tool.poetry.dependencies]
python = "^3.8"

pyyaml = "5.3"
kubernetes = "17.17.0"
inflection = "0.5.1"

[tool.poetry.dev-dependencies]
pytest = "6"
flake8 = "3.8"
black = "20.8b1"
